Add tests for fitting the GAM.
One test tests if the fitting procedure gives reasonable results
for constant data.

The other test compares our implementation with the tradeseq
implementation. For this, a wrapper is added that calls
tradeseq. But it only tests the placement of the knots and the
offset values. The comparison is done on the data from the
tradeseq tutorial.

See also: #13