Add method for evaluating diffent choices of n_knots.
Method that evaluates different choices for n_knots by fitting GAMs
on a subset of the genes for all given options for n_knots.
The different GAMs are evaluated by comparing the AICs.
Additionally, there is the possibility to plot different statistics
of the AIC.

Adds method returning the AIC of a GAM